wii: block driver for mem2 ram
Add support for accessing the unused part of the MEM2 RAM of the
Nintendo Wii video game console as a block device.

The MEM2 RAM is a second block of 64MB of DRAM starting at 0x10000000,
and thus not contiguous to the first block of 24MB of 1T-SRAM.
The default operating system running on the starlet co-processor uses
the top 12MB of MEM2, leaving ~52MB for free use.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>