b43: add config option to force QoS off
The b43 driver includes a capability mechanism that open source firmwares
(like OpenFWWF) can use to inform the driver about supported
characteristics.

The OpenFWWF firmware doesn't support yet QoS and reflects it correctly
via its capabilities word.

The b43 driver by default starts with QoS enabled (unless modparam_qos is
set to 0), but if the capabilities word of a firmware informs that QoS is
not supported then it tries to disable QoS.

Unfortunately, due to an unidentified bug, this automatic disabling of
QoS doesn't work. It works, though, if QoS is disabled from the very
beginning.

This patch adds a config option to allow disabling QoS from the beginning.
This is the only way to workaround the described problem when building the
b43 driver in-kernel, as in that case modparam_qos can't be changed.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>